+++ src/App.js
@@ -1,29 +1,13 @@
 import React from 'react';
-const todoList = [
-  {
-    id:1,
-     title: "Car Service"
-  },
-
-  {
-    id:2,
-    title: "School Conference"
-  },
-  {
-    id:3,
-    title: "Dental Appointment"
-  }
-
-];
+import TodoList from './components/TodoList'
+import AddTodoForm from './components/AddTodoForm'
 function App() {
   return (
     <>
      <h1>To Do List</h1>
-     <ul>
-       {todoList.map(function(item){
-         return <li key={item.objectID} >{item.title}</li>
-       })}
-     </ul>
+     <AddTodoForm/>
+     <TodoList/>
+     
     </>
   );
 }
